+++ bolt/__init__.py
@@ -5,6 +5,7 @@ from .problems.data_mixture import (
     DMCurriculumMO,
 )
 from .problems.hpo import HPO, HPOMultiFidelityModel, HPOMultiFidelityToken
+from .problems.parallelism_config import PCO16, PCO32, PCO64
 from .problems.prompt_opt import PO128, PO256, PO512, PO768
 
 __all__ = [
@@ -19,5 +20,8 @@ __all__ = [
     "PO256",
     "PO512",
     "PO768",
+    "PCO16",
+    "PCO32",
+    "PCO64",
 ]
-__version__ = "0.2.0.dev0"
+__version__ = "0.2.0.dev1"